daemon: Don't auto disable session affinity
This commit doesn't auto disable the session affinity feature when KPR
is disabled.

The session affinity doesn't depend on KPR for a ClusterIP service
translation in a pod netns implemented in bpf_lxc (the host netns case
should be covered by kube-proxy's rules). Therefore, we should allow
users to choose whether the feature should be enabled.

Note, that we don't enable the feature by default, as it might increase
bpf_lxc complexity.

Signed-off-by: Martynas Pumputis <m@lambda.lt>